This controller not uses resources of nut-tree-fork/nut-js package
